requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==0.12.3
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2-binary==2.7.5
requests==2.20.1
urllib3==1.24.1
Werkzeug==0.14.1
